feat: 패키지 위치 바꿨고 필드에 USer 객체를 갖도록 변경함
