+++ backend/src/main/java/com/example/demo/domain/user/service/CustomUserDetailsService.java
@@ -2,7 +2,7 @@ package com.example.demo.domain.user.service;
 
 import com.example.demo.domain.user.entity.User;
 import com.example.demo.domain.user.repository.UserRepository;
-import com.example.demo.global.security.UserPrincipal;
+import com.example.demo.domain.user.dto.UserPrincipal;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.userdetails.UserDetails;
